Move the guides to the processors directory
