+++ docs/api/guide.md
@@ -1,1 +1,1 @@
-::: outlines.fsm.guide
+::: outlines.processors.guide
