+++ outlines/processors/guide.py
