+++ tests/generate/conftest.py
@@ -9,13 +9,13 @@ def temp_cache_dir():
     import tempfile
 
     import outlines.caching
-    import outlines.fsm.guide
+    import outlines.processors.guide
 
     with tempfile.TemporaryDirectory() as tempdir:
         os.environ["OUTLINES_CACHE_DIR"] = tempdir
         outlines.caching.get_cache.cache_clear()
         reload(outlines)
-        reload(outlines.fsm.guide)
+        reload(outlines.processors.guide)
         cache_status = outlines.caching._caching_enabled
         try:
             outlines.caching._caching_enabled = True
